+++ dorc/if_run.py
@@ -13,6 +13,7 @@ if __name__ == '__main__':
     parser.add_argument("data_dir", type=str)
     parser.add_argument("gmods_dir", type=str)
     parser.add_argument("gdata_dir", type=str)
+    parser.add_argument("daemon_url", type=str)
     parser.add_argument("--config-overrides", type=str, default="False")
     args, _ = parser.parse_known_args()
     hostname = args.hostname
@@ -29,4 +30,4 @@ if __name__ == '__main__':
     # sys.path.append(os.path.join(data_dir, "..", ".."))
     sys.argv = [sys.argv[0]]
     FlaskInterface(hostname, port, data_dir, args.gmods_dir, args.gdata_dir,
-                   config_overrides=config)
+                   args.daemon_url, config_overrides=config)